requirements.txt
google-genai>=1.18.0
pydantic>=2.7.0
pandas>=2.2.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
numpy>=1.26.0
